requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cmake==4.0.0
contourpy==1.3.2
cycler==0.12.1
faiss-cpu==1.7.4
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.41
huggingface-hub==0.14.1
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.6
joblib==1.3.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lit==18.1.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.37.1
networkx==3.4.2
nltk==3.9.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.2
pandas==2.0.3
Pillow==9.5.0
protobuf==4.25.7
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.1
Pympler==1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
rich==13.9.4
rpds-py==0.24.0
scikit-learn==1.3.0
scipy==1.15.2
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
streamlit==1.25.0
sympy==1.14.0
tenacity==8.5.0
textblob==0.17.1
threadpoolctl==3.6.0
tokenizers==0.13.3
toml==0.10.2
torch==2.0.1
torchvision==0.15.2
tornado==6.4.2
tqdm==4.65.0
transformers==4.29.2
triton==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==4.3.1
urllib3==2.4.0
validators==0.35.0
watchdog==6.0.0
wordcloud==1.9.2
zipp==3.21.0
